refactor: destinify directory structure
