+++ frontend/App/scripts/shared/user-interaction.js
